feat: footnotes styling
