+++ src/services/graph.server.ts
@@ -14,7 +14,7 @@ import {
 import createLayout from 'ngraph.forcelayout'
 import createGraph from 'ngraph.graph'
 import { cached } from './cache.server.js'
-import { getPosts } from './posts.server.js'
+import { getPosts } from './posts/index.server.js'
 
 export const getGraph = cached(async (): Promise<Graph> => {
   const posts = await getPosts()
