+++ src/services/posts/types.ts
@@ -1,0 +1,18 @@
+export type Post = MarkdownPost | HrefPost
+
+export type MarkdownPost = BasePost & {
+  type: `markdown`
+  minutesToRead: number
+  content: string
+}
+
+export type HrefPost = BasePost & { type: `href`; href: string }
+
+type BasePost = {
+  id: string
+  title: string
+  tags: Set<string>
+  timestamp: Date
+  references: Map<string, Set<string>>
+  referencedBy: Set<string>
+}
