+++ src/styles/tailwind.css
@@ -17,24 +17,6 @@
   code {
     font-variant-ligatures: no-common-ligatures;
   }
-
-  h1,
-  h2,
-  h3,
-  h4,
-  h5,
-  h6 {
-    @apply relative;
-  }
-
-  :is(h1, h2, h3, h4, h5, h6)[id] a[href^='#'] {
-    @apply absolute top-1/2 h-6 w-6 -translate-x-7 -translate-y-1/2 opacity-0;
-  }
-
-  :is(h1, h2, h3, h4, h5, h6)[id]:hover a[href^='#'],
-  :is(h1, h2, h3, h4, h5, h6)[id] a[href^='#']:focus {
-    @apply opacity-100;
-  }
 }
 
 @layer utilities {
